Mejorando script de backup de archivos

+++ Backup_files.py
@@ -1,26 +1,31 @@
 import subprocess,time, os, sys
 
-sufijo=time.strftime("%Y%m%d-%I%M")
-###############CONFIGURACION########################
-proyecto=sys.argv[1]
-org=sys.argv[2]
-bckp=sys.argv[3]
-####################################################
-archivo=bckp+'/'+proyecto+'-'+sufijo+'.tgz'
-#Backup de archivos
-subprocess.call(['tar','-czvf',archivo,'--exclude=*.git','--exclude=*.mysql','--exclude=*.gz','--exclude=*.info',org])
+def backup(archivo,ruta,destino,opcion):
+    if opcion=='generate':
+        fecha=time.strftime("%Y%m%d-%I%M")
+        bckp=destino+'/'+archivo+'-'+fecha+'.tgz'
+        os.system('cd '+ruta)
+        subprocess.call(['tar','-czvf',bckp,'--exclude=*.git','--exclude=*.mysql','--exclude=*.gz','--exclude=*.info',archivo])
+    else:
+        #Limpia Backup
+        year=time.strftime("%Y")
+        month=time.strftime("%m")
+        if month=='01':
+            last_month=12
+            year= int(year)-1
+            year= str(year)
+        else:
+            last_month= int(month)-1
 
-#Limpia Backup
-year=time.strftime("%Y")
-month=time.strftime("%m")
-if month=='01':
-    last_month=12
-    year= int(year)-1
-    year= str(year)
-else:
-    last_month= int(month)-1
+        n="{:02d}".format(last_month)
+        m= str(n)
+        p=destino+'/'+archivo+'-'+year+m+'*'
+        os.system('rm -f '+p)
 
-n="{:02d}".format(last_month)
-m= str(n)
-p=bckp+'/'+proyecto+'-'+year+m+'*'
-os.system('rm -f '+p)
+archivo=sys.argv[1]
+ruta=sys.argv[2]
+destino=sys.argv[3]
+opcion=sys.argv[4]
+#Ejemplo
+#backup('pp.txt','/home/lmayta/Phyton','/home/lmayta')
+backup(archivo,ruta,destino,opcion)
